pyproject.toml
[project]
name = "bagusanmana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.56.0",
    "pillow>=12.0.0",
]
